On-behalf-of: SAP Michal Krzyz <michal.krzyz@sap.com>

+++ internal/util/user.go
@@ -3,4 +3,4 @@
 
 package util
 
-const SystemUserId = 1
+const SystemUserId = int64(1)
